requirements.txt
streamlit==1.28.0
pinecone-client==3.2.2
sentence-transformers==2.2.2
rank-bm25==0.2.2
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
scikit-learn==1.3.0
nltk==3.8.1
tqdm==4.65.0